fix: harden sentiment insights integration
