+++ __tests__/adanos.actions.test.ts
@@ -1,10 +1,19 @@
-import { describe, expect, it } from 'vitest';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 
+import {
+    getStockSentimentInsights,
+} from '@/lib/actions/adanos.actions';
 import {
     buildStockSentimentInsights,
     getSourceAlignment,
     normalizeSourceInsight,
-} from '@/lib/actions/adanos.actions';
+} from '@/lib/actions/adanos.helpers';
+
+afterEach(() => {
+    vi.restoreAllMocks();
+    delete process.env.ADANOS_API_KEY;
+    delete process.env.ADANOS_API_BASE_URL;
+});
 
 describe('normalizeSourceInsight', () => {
     it('maps source-specific metrics for mentions and trades', () => {
@@ -120,5 +129,84 @@ describe('buildStockSentimentInsights', () => {
 
     it('returns null when no sources have usable data', () => {
         expect(buildStockSentimentInsights('MSFT', [null, null])).toBeNull();
+    });
+});
+
+describe('getStockSentimentInsights', () => {
+    it('returns a parsed result when compare data matches the requested ticker', async () => {
+        process.env.ADANOS_API_KEY = 'test-key';
+        vi.spyOn(global, 'fetch').mockImplementation(async (input) => {
+            const url = String(input);
+
+            if (url.includes('/reddit/')) {
+                return new Response(
+                    JSON.stringify({
+                        stocks: [{ ticker: 'TSLA', company_name: 'Tesla, Inc.', buzz_score: 80, bullish_pct: 40, trend: 'rising', mentions: 10 }],
+                    }),
+                    { status: 200 },
+                );
+            }
+
+            if (url.includes('/x/')) {
+                return new Response(
+                    JSON.stringify({
+                        stocks: [{ ticker: 'TSLA', company_name: 'Tesla, Inc.', buzz_score: 90, bullish_pct: 60, trend: 'falling', mentions: 20 }],
+                    }),
+                    { status: 200 },
+                );
+            }
+
+            return new Response(JSON.stringify({ stocks: [] }), { status: 404 });
+        });
+
+        const insight = await getStockSentimentInsights('TSLA');
+
+        expect(insight).toMatchObject({
+            symbol: 'TSLA',
+            companyName: 'Tesla, Inc.',
+            averageBuzz: 85,
+            bullishAverage: 50,
+            availableSources: 2,
+        });
+        expect(insight?.sources).toHaveLength(2);
+    });
+
+    it('returns null when the remote source returns 404 for all sources', async () => {
+        process.env.ADANOS_API_KEY = 'test-key';
+        vi.spyOn(global, 'fetch').mockResolvedValue(new Response(null, { status: 404 }));
+
+        await expect(getStockSentimentInsights('TSLA')).resolves.toBeNull();
+    });
+
+    it('returns null when the remote payload contains a different ticker only', async () => {
+        process.env.ADANOS_API_KEY = 'test-key';
+        vi.spyOn(global, 'fetch').mockResolvedValue(
+            new Response(
+                JSON.stringify({
+                    stocks: [{ ticker: 'MSFT', company_name: 'Microsoft Corporation', buzz_score: 70, bullish_pct: 55, trend: 'stable', mentions: 30 }],
+                }),
+                { status: 200 },
+            ),
+        );
+
+        await expect(getStockSentimentInsights('TSLA')).resolves.toBeNull();
+    });
+
+    it('returns null when the response body is invalid json', async () => {
+        process.env.ADANOS_API_KEY = 'test-key';
+        vi.spyOn(global, 'fetch').mockResolvedValue({
+            status: 200,
+            ok: true,
+            json: vi.fn().mockRejectedValue(new Error('invalid json')),
+        } as unknown as Response);
+
+        await expect(getStockSentimentInsights('TSLA')).resolves.toBeNull();
+    });
+
+    it('returns null when fetch fails', async () => {
+        process.env.ADANOS_API_KEY = 'test-key';
+        vi.spyOn(global, 'fetch').mockRejectedValue(new Error('network failed'));
+
+        await expect(getStockSentimentInsights('TSLA')).resolves.toBeNull();
     });
 });
