+++ components/stocks/StockSentimentCard.tsx
@@ -1,4 +1,4 @@
-import type { StockSentimentInsights } from '@/lib/actions/adanos.actions';
+import type { StockSentimentInsights } from '@/lib/actions/adanos.helpers';
 
 interface StockSentimentCardProps {
     insight: StockSentimentInsights | null;
@@ -27,6 +27,9 @@ function getAlignmentClasses(alignment: string): string {
     if (alignment === 'Bullish alignment') return 'text-emerald-400';
     if (alignment === 'Bearish alignment' || alignment === 'Wide divergence') return 'text-rose-400';
     if (alignment === 'Tight alignment') return 'text-blue-300';
+    if (alignment === 'Mixed') return 'text-amber-300';
+    if (alignment === 'Single-source view') return 'text-slate-300';
+    if (alignment === 'No sentiment mix') return 'text-zinc-400';
     return 'text-gray-300';
 }
